`svm-hash` crate.
