--- crates/svm-common/src/ffi.rs
@@ -1,1 +1,0 @@
-
