+++ crates/svm-common/src/lib.rs
@@ -5,12 +5,5 @@
 
 //! `svm-common` crate groups common shared code between the other `SVM` crates
 
-mod default_key_hasher;
-mod ffi;
-mod key_hasher;
-
 /// Common Formatters
 pub mod fmt;
-
-pub use default_key_hasher::DefaultKeyHasher;
-pub use key_hasher::KeyHasher;
