+++ crates/svm-runtime-c-api/Cargo.toml
@@ -15,9 +15,7 @@ crate-type = ["cdylib", "staticlib", "rlib"]
 
 [dependencies]
 log = "0.4"
-byteorder = "1.3.2"
 svm-runtime = { path = "../svm-runtime", default-features = false }
-svm-common = { path = "../svm-common" }
 svm-kv = { path = "../svm-kv", default-features = false }
 svm-ffi = { path = "../svm-ffi" }
 svm-types = { path = "../svm-types" }
