+++ crates/svm-runtime/Cargo.toml
@@ -18,6 +18,7 @@ byteorder = "1.3.2"
 lazy_static = "1.4.0"
 wat = "1.0"
 svm-ffi = { path = "../svm-ffi" }
+svm-hash = { path = "../svm-hash" }
 svm-common = { path = "../svm-common" }
 svm-types = { path = "../svm-types" }
 svm-layout = { path = "../svm-layout" }
